refactor(apis_entities, apis_relations): move filters and methods
Move the Property and Triple related filters and methods to the
`apis_relations.filters` module instead of keeping them in the
`apis_entities.filtersets` module. This makes the whole module more
readable and moves the filters and methods closer to the models they are
working with. Additionaly, this makes it possible to only import the
fitlers and methods if the `apis_relations` app is actually part of the
INSTALLED_APPS list.
